# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/material": {
            "post": {
                "description": "Create a new material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new material",
                "operationId": "create-material",
                "parameters": [
                    {
                        "description": "Create Material Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/material/delete": {
            "delete": {
                "description": "Delete a material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a material",
                "operationId": "delete-material",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Material ID",
                        "name": "material_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Course ID",
                        "name": "course_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/material/edit": {
            "put": {
                "description": "Update a material",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a material",
                "operationId": "update-material",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Material ID",
                        "name": "material_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Material Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpdateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/material/get": {
            "get": {
                "description": "Get Material of certain course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Material",
                "parameters": [
                    {
                        "description": "Get Material of certain course",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.DeleteMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/material/list": {
            "get": {
                "description": "List Material of certain course",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List Material",
                "parameters": [
                    {
                        "description": "List Material of certain course",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ListMaterialsByCourse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CreateMaterialRequest": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                },
                "material_file": {
                    "type": "string"
                },
                "order_number": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "api.DeleteMaterialRequest": {
            "type": "object",
            "properties": {
                "material_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.ListMaterialsByCourse": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                }
            }
        },
...
        "api.UpdateMaterialRequest": {
            "type": "object",
            "properties": {
                "course_id": {
                    "type": "integer"
                },
                "material_file": {
                    "type": "string"
                },
                "material_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}